reformat code

+++ src/main/java/net/fap/beecloud/math/BeeCloudMath.java
@@ -2,14 +2,15 @@ package net.fap.beecloud.math;
 
 public class BeeCloudMath {
 
-    public BeeCloudMath(){}
+	public BeeCloudMath() {
+	}
 
-    public static double round(double d) {
-        return round(d, 0);
-    }
+	public static double round(double d) {
+		return round(d, 0);
+	}
 
-    public static double round(double d, int precision) {
-        return (double)Math.round(d * Math.pow(10.0D, (double)precision)) / Math.pow(10.0D, (double)precision);
-    }
+	public static double round(double d, int precision) {
+		return (double) Math.round(d * Math.pow(10.0D, precision)) / Math.pow(10.0D, precision);
+	}
 
 }
